+++ spec/services/zammad_customer_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe ZammadCustomer do
         let!(:user) { create(:user, email: "soukalina@gmail.com") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(email: "soukalina@gmail.com", phone: nil)
+          zammad_customer = ZammadCustomer.new(email: "soukalina@gmail.com", phone: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Usager trouvé avec l'email soukalina@gmail.com"
           expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/users/#{user.id}"
@@ -17,7 +17,7 @@ RSpec.describe ZammadCustomer do
         let(:args) { {} }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(email: "agent@example.com", phone: nil)
+          zammad_customer = ZammadCustomer.new(email: "agent@example.com", phone: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Agent trouvé avec l'email agent@example.com"
           expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/agents/#{agent.id}"
@@ -28,7 +28,7 @@ RSpec.describe ZammadCustomer do
         let!(:user) { create(:user, phone_number: "0612345678", phone_number_formatted: "+33612345678") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(phone: "0612345678", email: nil)
+          zammad_customer = ZammadCustomer.new(phone: "0612345678", email: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Usager trouvé avec le numéro de téléphone formatté +33612345678"
           expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/users/#{user.id}"
@@ -37,7 +37,7 @@ RSpec.describe ZammadCustomer do
 
       context "aucun usager ni agent trouvé" do
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(email: "unknown@example.com", phone: "013456789")
+          zammad_customer = ZammadCustomer.new(email: "unknown@example.com", phone: "013456789")
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Aucun usager ni agent trouvé"
           expect(zammad_customer.super_admin_url).to be_nil
@@ -49,7 +49,7 @@ RSpec.describe ZammadCustomer do
         let!(:user_by_phone) { create(:user, phone_number: "0612345678") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(email: "test@example.com", phone: "0612345678")
+          zammad_customer = ZammadCustomer.new(email: "test@example.com", phone: "0612345678")
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Usager trouvé avec l'email test@example.com"
           expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/users/#{user_by_email.id}"
@@ -61,7 +61,7 @@ RSpec.describe ZammadCustomer do
         let!(:user) { create(:user, phone_number: "06.12-34.56.78", phone_number_formatted: "+33612345678") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(phone: "06 12 34 56 78", email: nil)
+          zammad_customer = ZammadCustomer.new(phone: "06 12 34 56 78", email: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Usager trouvé avec le numéro de téléphone formatté +33612345678"
           expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/users/#{user.id}"
@@ -73,7 +73,7 @@ RSpec.describe ZammadCustomer do
         let!(:user2) { create(:user, phone_number: "06.12-34.56.78", phone_number_formatted: "+33612345678kdkd") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(phone: "06 12 34 56 78", email: nil)
+          zammad_customer = ZammadCustomer.new(phone: "06 12 34 56 78", email: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Plusieurs usagers trouvés avec le numéro de téléphone formatté +33612345678"
           expect(zammad_customer.super_admin_url).to be_nil
@@ -85,7 +85,7 @@ RSpec.describe ZammadCustomer do
         let!(:user2) { build(:user, phone_number: "01 23 44").tap { _1.save!(validate: false) } }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(phone: "01 23 44", email: nil)
+          zammad_customer = ZammadCustomer.new(phone: "01 23 44", email: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Plusieurs usagers trouvés avec le numéro de téléphone 01 23 44"
           expect(zammad_customer.super_admin_url).to be_nil
@@ -96,7 +96,7 @@ RSpec.describe ZammadCustomer do
         let!(:user) { create(:user, phone_number: "0123456789") }
 
         specify do
-          zammad_customer = ZammadCustomer::Attributes.new(phone: "123", email: nil)
+          zammad_customer = ZammadCustomer.new(phone: "123", email: nil)
           described_class.new(zammad_customer).run
           expect(zammad_customer.note).to eq "Aucun usager ni agent trouvé"
           expect(zammad_customer.super_admin_url).to be_nil
@@ -108,7 +108,7 @@ RSpec.describe ZammadCustomer do
       let!(:agent) { create(:agent, email: "agent@example.com") }
 
       specify do
-        zammad_customer = ZammadCustomer::Attributes.new(phone: "123", email: nil)
+        zammad_customer = ZammadCustomer.new(phone: "123", email: nil)
         described_class.new(zammad_customer).augment_with_agent(agent)
         expect(zammad_customer.note).to be_blank
         expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/agents/#{agent.id}"
@@ -119,7 +119,7 @@ RSpec.describe ZammadCustomer do
       let!(:user) { create(:user, email: "user@example.com") }
 
       specify do
-        zammad_customer = ZammadCustomer::Attributes.new(phone: "123", email: nil)
+        zammad_customer = ZammadCustomer.new(phone: "123", email: nil)
         described_class.new(zammad_customer).augment_with_user(user)
         expect(zammad_customer.note).to be_blank
         expect(zammad_customer.super_admin_url).to eq "http://www.rdv-mairie-test.localhost/super_admins/users/#{user.id}"
